fix: isolate visual build and align OpenCV ABI

+++ src/lvi_sam/src/visual_odometry/visual_feature/camera_models/EquidistantCamera.cc
@@ -668,7 +668,7 @@ EquidistantCamera::setParameters(const EquidistantCamera::Parameters& parameters
 void
 EquidistantCamera::readParameters(const std::vector<double>& parameterVec)
 {
-    if (parameterVec.size() != parameterCount())
+    if (parameterVec.size() != static_cast<std::size_t>(parameterCount()))
     {
         return;
     }
